# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Daemon Host Panel
  version: 0.1.0
paths:
...
  /api/participant/debate/pick-side:
    post:
      tags:
      - debate
      summary: Pick Side
      description: Participant picks a side (for/against).
      operationId: pick_side_api_participant_debate_pick_side_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PickSideRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: debate
  /api/participant/debate/argument:
    post:
      tags:
      - debate
      summary: Submit Argument
      description: Participant submits a debate argument.
      operationId: submit_argument_api_participant_debate_argument_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArgumentRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: debate
  /api/participant/debate/upvote:
    post:
      tags:
      - debate
      summary: Upvote Argument
      description: Participant upvotes a debate argument.
      operationId: upvote_argument_api_participant_debate_upvote_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpvoteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: debate
...
  /api/{session_id}/host/debate:
    post:
      tags:
      - debate
      summary: Launch Debate
      description: Host launches a debate with a statement.
      operationId: launch_debate_api__session_id__host_debate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchDebateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: debate
...
  /api/{session_id}/host/debate/phase:
    post:
      tags:
      - debate
      summary: Advance Phase
      description: Host advances the debate to a specific phase.
      operationId: advance_phase_api__session_id__host_debate_phase_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvancePhaseRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkPhaseResponse'
      x-feature: debate
  /api/{session_id}/host/debate/first-side:
    post:
      tags:
      - debate
      summary: Set First Side
      description: Host picks which side speaks first in live debate.
      operationId: set_first_side_api__session_id__host_debate_first_side_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetFirstSideRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: debate
  /api/{session_id}/host/debate/round-timer:
    post:
      tags:
      - debate
      summary: Start Round Timer
      description: Host starts a timed round.
      operationId: start_round_timer_api__session_id__host_debate_round_timer_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoundTimerRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: debate
...
  /api/{session_id}/host/debate/ai-result:
    post:
      tags:
      - debate
      summary: Receive Ai Result
      description: Manual/skip AI result — host posts AI cleanup results directly.
      operationId: receive_ai_result_api__session_id__host_debate_ai_result_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiResultRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: debate
...
components:
  schemas:
    AdvancePhaseRequest:
      properties:
        phase:
          type: string
          title: Phase
      type: object
      required:
      - phase
      title: AdvancePhaseRequest
    AiResultRequest:
      properties:
        merges:
          items:
            $ref: '#/components/schemas/DebateAiMerge'
          type: array
          title: Merges
          default: []
        cleaned:
          items:
            $ref: '#/components/schemas/DebateAiCleaned'
          type: array
          title: Cleaned
          default: []
        new_arguments:
          items:
            $ref: '#/components/schemas/DebateAiNewArgument'
          type: array
          title: New Arguments
          default: []
      type: object
      title: AiResultRequest
    ArgumentRequest:
      properties:
        text:
          type: string
          title: Text
      type: object
      required:
      - text
      title: ArgumentRequest
...
    DebateAiCleaned:
      properties:
        id:
          type: string
          title: Id
        text:
          type: string
          title: Text
      type: object
      required:
      - id
      - text
      title: DebateAiCleaned
    DebateAiMerge:
      properties:
        keep_id:
          type: string
          title: Keep Id
        remove_ids:
          items:
            type: string
          type: array
          title: Remove Ids
          default: []
      type: object
      required:
      - keep_id
      title: DebateAiMerge
    DebateAiNewArgument:
      properties:
        side:
          type: string
          title: Side
        text:
          type: string
          title: Text
      type: object
      required:
      - side
      - text
      title: DebateAiNewArgument
...
    LaunchDebateRequest:
      properties:
        statement:
          type: string
          title: Statement
      type: object
      required:
      - statement
      title: LaunchDebateRequest
...
    OkPhaseResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        phase:
          type: string
          title: Phase
      type: object
      required:
      - phase
      title: OkPhaseResponse
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
...
    PickSideRequest:
      properties:
        side:
          type: string
          title: Side
      type: object
      required:
      - side
      title: PickSideRequest
...
    RoundTimerRequest:
      properties:
        round_index:
          type: integer
          title: Round Index
        seconds:
          type: integer
          title: Seconds
      type: object
      required:
      - round_index
      - seconds
      title: RoundTimerRequest
...
    SetFirstSideRequest:
      properties:
        side:
          type: string
          title: Side
      type: object
      required:
      - side
      title: SetFirstSideRequest
...
    UpvoteRequest:
      properties:
        argument_id:
          type: string
          title: Argument Id
      type: object
      required:
      - argument_id
      title: UpvoteRequest
...